차트 크롤링 구현(80%) (01.20)
80%정도 구현
테스트는 안해봤습니다.

+++ Code/backend/berry_comment/src/main/java/com/berry_comment/dto/CrawlInfoDto.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Builder
 public class CrawlInfoDto {
     //곡 아이디
-    private int songId;
+    private Long songId;
 
     //곡 이름
     private String song;
