+++ Code/backend/berry_comment/src/main/java/com/berry_comment/repository/SongRepository.java
@@ -3,5 +3,6 @@ package com.berry_comment.repository;
 import com.berry_comment.entity.Song;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SongRepository extends JpaRepository<Song,Integer> {
+public interface SongRepository extends JpaRepository<Song,Long> {
+
 }
